% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCCL.R
\docType{package}
\name{iCCL-package}
\alias{iCCL-package}
\title{iCCL: interactive Comparative Clustering Loop}
\description{
Objective PCA dimensionality estimation and multi-dimension / multi-resolution
Seurat clustering with UMAP and t-SNE embeddings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ari-abb/iCCL}
  \item Report bugs at \url{https://github.com/ari-abb/iCCL/issues}
}

}
\author{
Arian Abbasi \email{abbasi@hhu.de}
}
\keyword{internal}
