% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCCL1.R
\docType{package}
\name{iCCL1-package}
\alias{iCCL1}
\alias{iCCL1-package}
\title{iCCL1: interactive Comparative Clustering Loop}
\description{
Objective PCA dimensionality estimation and multi-dimension / multi-resolution
Seurat clustering with UMAP and t-SNE embeddings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ari7cr/iCCL1}
  \item Report bugs at \url{https://github.com/ari7cr/iCCL1/issues}
}

}
\author{
Arian Abbasi \email{abbasi@hhu.de}
}
\keyword{internal}
